+++ core/vtk/ttkRangePolygon/ttkRangePolygon.cpp
@@ -137,7 +137,7 @@ int ttkRangePolygon::processTriangles(vtkUnstructuredGrid *input,
       });
     }
     
-    for(int i = 0; i < (int) output->GetPointData()->GetNumberOfArrays(); i++){
+    for(int i = 0; i < output->GetPointData()->GetNumberOfArrays(); i++){
       vtkDataArray *field = output->GetPointData()->GetArray(i);
       
       switch(field->GetDataType()){
